backend/requirements.txt
# Core Framework (Updated for pydantic v2 + Python 3.12 compatibility)
langchain==0.3.13
langchain-community==0.3.13
langchain-core>=0.3.29,<0.4.0
langchain-text-splitters==0.3.5

# LLM & Embeddings
anthropic>=0.40.0
sentence-transformers>=2.7.0
langchain-huggingface==0.3.1

# Vector Database
chromadb>=0.4.22
langchain-chroma==0.2.6

# Web Framework (Week 2)
fastapi>=0.119.0
uvicorn[standard]==0.27.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
slowapi==0.1.9
python-multipart==0.0.6
starlette>=0.48.0

# Document Processing
pypdf==3.17.4
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2

# Streamlit UI (Week 1)
streamlit==1.29.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-watch==4.2.0
httpx>=0.28.1
black==23.12.1
flake8==7.0.0
mypy==1.8.0
